migrate all vcsurl usages to the new implementation
